+++ dwara-videopub/src/main/java/org/ishafoundation/videopub/transcoding/ffmpeg/audio/Audio_LowResolution_Transcoding_TaskExecutor.java
@@ -1,0 +1,96 @@
+package org.ishafoundation.videopub.transcoding.ffmpeg.audio;
+
+import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.FilenameUtils;
+import org.ishafoundation.dwaraapi.commandline.local.CommandLineExecutionResponse;
+import org.ishafoundation.dwaraapi.process.IProcessingTask;
+import org.ishafoundation.dwaraapi.process.LogicalFile;
+import org.ishafoundation.dwaraapi.process.ProcessingtaskResponse;
+import org.ishafoundation.dwaraapi.process.request.ProcessContext;
+import org.ishafoundation.videopub.transcoding.ffmpeg.MediaTask;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Component;
+
+@Component("audio-proxy-low-gen")
+@Primary
+@Profile({ "!dev & !stage" })
+public class Audio_LowResolution_Transcoding_TaskExecutor extends MediaTask implements IProcessingTask{
+
+	private static final Logger logger = LoggerFactory.getLogger(Audio_LowResolution_Transcoding_TaskExecutor.class);
+
+	@Override
+	public ProcessingtaskResponse execute(ProcessContext processContext) throws Exception {
+		
+		String taskName = processContext.getJob().getProcessingtaskId();
+		LogicalFile logicalFile = processContext.getLogicalFile();
+		Integer fileId = processContext.getFile().getId();
+		String destinationDirPath = processContext.getOutputDestinationDirPath();
+		
+		if(logger.isTraceEnabled()) {
+			logger.trace("taskName " + taskName);
+//			logger.trace("inputArtifactName " + inputArtifactName); // V22205_Test_5D-Camera_Mahabharat_Day7-Morning_Isha-Samskriti-Singing_AYA_17-Feb-12
+//			logger.trace("outputArtifactName " + outputArtifactName); // VL22205_Test_5D-Camera_Mahabharat_Day7-Morning_Isha-Samskriti-Singing_AYA_17-Feb-12
+			logger.trace("fileId " + fileId);
+//			logger.trace("domain " + domain.name()); 
+			logger.trace("logicalFile " + logicalFile.getAbsolutePath()); // /data/ingested/V22205_Test_5D-Camera_Mahabharat_Day7-Morning_Isha-Samskriti-Singing_AYA_17-Feb-12/DCIM/100EOS5D/MVI_5594.MOV
+//			logger.trace("category " + category); // public
+			logger.trace("destinationDirPath " + destinationDirPath); // /data/transcoded/public/VL22205_Test_5D-Camera_Mahabharat_Day7-Morning_Isha-Samskriti-Singing_AYA_17-Feb-12/DCIM/100EOS5D
+		}
+		
+		String sourceFilePathname = logicalFile.getAbsolutePath();
+		
+		String identifierSuffix = "";//mediaFileId+"";//  +  "-" + VideoProcessingSteps.PROXY_GENERATION.toString();
+		String containerName = identifierSuffix;	
+		
+		FileUtils.forceMkdir(new File(destinationDirPath));
+	
+		String fileName = FilenameUtils.getBaseName(sourceFilePathname);
+		String proxyTargetLocation = destinationDirPath + File.separator + fileName + ".mp3";		
+		
+		/*************** PROXY GENERATION ***************/
+		long proxyStartTime = System.currentTimeMillis();
+		logger.info("Proxy Generation start for " + containerName + " - targetLocation is - " + proxyTargetLocation);
+	
+		// Doing this command creation and execution in 2 steps so that the process can be referenced in memory and so if cancel command for a specific medialibrary is issued the specific process(es) can be destroyed/killed referencing this...
+		// mapping only for proxy generation commands which are slightly heavy and time consuming than the thumbnail and metadata extraction...
+		List<String> proxyGenerationCommandParamsList = getProxyGenCommand(sourceFilePathname, proxyTargetLocation);
+		CommandLineExecutionResponse proxyCommandLineExecutionResponse = createProcessAndExecuteCommand(fileId+"~"+taskName , proxyGenerationCommandParamsList);
+		long proxyEndTime = System.currentTimeMillis();
+	
+		// TODO : better this...
+		ProcessingtaskResponse processingtaskResponse = new ProcessingtaskResponse();
+		processingtaskResponse.setDestinationPathname(proxyTargetLocation);
+		processingtaskResponse.setIsComplete(proxyCommandLineExecutionResponse.isComplete());
+		processingtaskResponse.setIsCancelled(proxyCommandLineExecutionResponse.isCancelled());
+		processingtaskResponse.setStdOutResponse(proxyCommandLineExecutionResponse.getStdOutResponse());
+		processingtaskResponse.setFailureReason(proxyCommandLineExecutionResponse.getFailureReason());
+
+		return processingtaskResponse;
+	}
+	
+
+	// ffmpeg -y -nostdin -i <<sourceFilePathname>> -b:a 64k <<proxyTargetLocation>> 
+	// ffmpeg -y -nostdin -i file.wav -b:a 64k file.mp3
+	// ffmpeg -y -nostdin -i "Tel Dub.wav" -b:a 64000 "output/Tel Dub.mp3" 
+	private List<String> getProxyGenCommand(String sourceFilePathname, String proxyTargetLocation) {
+		List<String> proxyGenerationCommandParamsList = new ArrayList<String>();
+		
+		proxyGenerationCommandParamsList.add("ffmpeg");
+		proxyGenerationCommandParamsList.add("-y");
+		proxyGenerationCommandParamsList.add("-nostdin");
+		proxyGenerationCommandParamsList.add("-i");
+		proxyGenerationCommandParamsList.add(sourceFilePathname);
+		proxyGenerationCommandParamsList.add("-b:a");
+		proxyGenerationCommandParamsList.add("64k");
+		proxyGenerationCommandParamsList.add(proxyTargetLocation);
+	
+		return proxyGenerationCommandParamsList;
+	}
+}