+++ dwara-commons/src/main/java/org/ishafoundation/dwaraapi/commandline/remote/scp/SecuredCopier.java
@@ -2,6 +2,7 @@ package org.ishafoundation.dwaraapi.commandline.remote.scp;
 
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
 
@@ -11,6 +12,7 @@ import org.springframework.stereotype.Component;
 
 import com.jcraft.jsch.Channel;
 import com.jcraft.jsch.ChannelExec;
+import com.jcraft.jsch.ChannelSftp;
 import com.jcraft.jsch.Session;
 
 /**
@@ -21,6 +23,124 @@ import com.jcraft.jsch.Session;
 public class SecuredCopier {
 	
 	static Logger logger = LoggerFactory.getLogger(SecuredCopier.class);
+	
+	public void copyFrom(Session jschSession, String remoteFilePath, String localFilePath) throws Exception{
+		 ChannelSftp channelSftp =  null;
+        try {
+        	Channel sftp = jschSession.openChannel("sftp");
+
+	        // 5 seconds timeout
+	        sftp.connect(5000);
+	
+	        channelSftp = (ChannelSftp) sftp;
+	
+	        // download file from remote server to local
+	        channelSftp.get(remoteFilePath, localFilePath);
+        }
+        finally {
+        	if(channelSftp != null)
+        		channelSftp.exit();	
+		}
+        
+	}
+	
+	public void copyFrom(Session session, String from, String to, String fileName) throws Exception {
+        from = from + File.separator + fileName;
+        String prefix = null;
+
+        if (new File(to).isDirectory()) {
+            prefix = to + File.separator;
+        }
+
+        // exec 'scp -f rfile' remotely
+        String command = "scp -f " + from;
+        Channel channel = session.openChannel("exec");
+        ((ChannelExec) channel).setCommand(command);
+
+        // get I/O streams for remote scp
+        OutputStream out = channel.getOutputStream();
+        InputStream in = channel.getInputStream();
+
+        channel.connect();
+
+        byte[] buf = new byte[1024];
+
+        // send '\0'
+        buf[0] = 0;
+        out.write(buf, 0, 1);
+        out.flush();
+
+        while (true) {
+            int c = checkAck(in);
+            if (c != 'C') {
+                break;
+            }
+
+            // read '0644 '
+            in.read(buf, 0, 5);
+
+            long filesize = 0L;
+            while (true) {
+                if (in.read(buf, 0, 1) < 0) {
+                    // error
+                    break;
+                }
+                if (buf[0] == ' ') break;
+                filesize = filesize * 10L + (long) (buf[0] - '0');
+            }
+
+            String file = null;
+            for (int i = 0; ; i++) {
+                in.read(buf, i, 1);
+                if (buf[i] == (byte) 0x0a) {
+                    file = new String(buf, 0, i);
+                    break;
+                }
+            }
+
+            logger.debug("file-size=" + filesize + ", file=" + file);
+
+            // send '\0'
+            buf[0] = 0;
+            out.write(buf, 0, 1);
+            out.flush();
+
+            // read a content of lfile
+            FileOutputStream fos = new FileOutputStream(prefix == null ? to : prefix + file);
+            int foo;
+            while (true) {
+                if (buf.length < filesize) foo = buf.length;
+                else foo = (int) filesize;
+                foo = in.read(buf, 0, foo);
+                if (foo < 0) {
+                    // error
+                    break;
+                }
+                fos.write(buf, 0, foo);
+                filesize -= foo;
+                if (filesize == 0L) break;
+            }
+
+            if (checkAck(in) != 0) {
+            	throw new Exception("\"C0644 filesize filename\" ack failed");
+            }
+
+            // send '\0'
+            buf[0] = 0;
+            out.write(buf, 0, 1);
+            out.flush();
+
+            try {
+                if (fos != null) fos.close();
+            } catch (Exception ex) {
+                
+            }
+        }
+
+        channel.disconnect();
+        session.disconnect();
+    }
+
 	
 	public void copyTo(Session session, String localFilePath, String remoteFilePath) throws Exception {
 		FileInputStream fis = null;